+++ script/tool/test/publish_plugin_command_test.dart
@@ -30,6 +30,8 @@ void main() {
   late CommandRunner<void> commandRunner;
   late MockStdin mockStdin;
   late FileSystem fileSystem;
+  // Map of package name to mock response.
+  late Map<String, Map<String, dynamic>> mockHttpResponses;
 
   void _createMockCredentialFile() {
     final String credentialPath = PublishPluginCommand.getCredentialPath();
@@ -41,8 +43,20 @@ void main() {
   setUp(() async {
     fileSystem = MemoryFileSystem();
     packagesDir = createPackagesDirectory(fileSystem: fileSystem);
-
     processRunner = TestProcessRunner();
+
+    mockHttpResponses = <String, Map<String, dynamic>>{};
+    final MockClient mockClient = MockClient((http.Request request) async {
+      final String packageName =
+          request.url.pathSegments.last.replaceAll('.json', '');
+      final Map<String, dynamic>? response = mockHttpResponses[packageName];
+      if (response != null) {
+        return http.Response(json.encode(response), 200);
+      }
+      // Default to simulating the plugin never having been published.
+      return http.Response('', 404);
+    });
+
     gitDir = MockGitDir();
     when(gitDir.path).thenReturn(packagesDir.parent.path);
     when(gitDir.runCommand(any, throwOnError: anyNamed('throwOnError')))
@@ -58,39 +72,16 @@ void main() {
 
     mockStdin = MockStdin();
     commandRunner = CommandRunner<void>('tester', '')
-      ..addCommand(PublishPluginCommand(packagesDir,
-          processRunner: processRunner, stdinput: mockStdin, gitDir: gitDir));
+      ..addCommand(PublishPluginCommand(
+        packagesDir,
+        processRunner: processRunner,
+        stdinput: mockStdin,
+        gitDir: gitDir,
+        httpClient: mockClient,
+      ));
   });
 
   group('Initial validation', () {
-    test('requires a package flag', () async {
-      Error? commandError;
-      final List<String> output = await runCapturingPrint(
-          commandRunner, <String>['publish-plugin'], errorHandler: (Error e) {
-        commandError = e;
-      });
-
-      expect(commandError, isA<ToolExit>());
-      expect(
-          output,
-          containsAllInOrder(<Matcher>[
-            contains('Must specify a package to publish.'),
-          ]));
-    });
-
-    test('requires an existing flag', () async {
-      Error? commandError;
-      final List<String> output = await runCapturingPrint(
-          commandRunner, <String>['publish-plugin', '--package', 'iamerror'],
-          errorHandler: (Error e) {
-        commandError = e;
-      });
-
-      expect(commandError, isA<ToolExit>());
-      expect(output,
-          containsAllInOrder(<Matcher>[contains('iamerror does not exist')]));
-    });
-
     test('refuses to proceed with dirty files', () async {
       final Directory pluginDir =
           createFakePlugin('foo', packagesDir, examples: <String>[]);
@@ -100,9 +91,11 @@ void main() {
       ];
 
       Error? commandError;
-      final List<String> output = await runCapturingPrint(
-          commandRunner, <String>['publish-plugin', '--package', 'foo'],
-          errorHandler: (Error e) {
+      final List<String> output =
+          await runCapturingPrint(commandRunner, <String>[
+        'publish-plugin',
+        '--packages=foo',
+      ], errorHandler: (Error e) {
         commandError = e;
       });
 
@@ -128,7 +121,7 @@ void main() {
 
       Error? commandError;
       final List<String> output = await runCapturingPrint(
-          commandRunner, <String>['publish-plugin', '--package', 'foo'],
+          commandRunner, <String>['publish-plugin', '--packages=foo'],
           errorHandler: (Error e) {
         commandError = e;
       });
@@ -145,24 +138,34 @@ void main() {
 
   group('Publishes package', () {
     test('while showing all output from pub publish to the user', () async {
-      createFakePlugin('foo', packagesDir, examples: <String>[]);
+      createFakePlugin('plugin1', packagesDir, examples: <String>[]);
+      createFakePlugin('plugin2', packagesDir, examples: <String>[]);
 
       processRunner.mockProcessesForExecutable[flutterCommand] = <io.Process>[
         MockProcess(
             stdout: 'Foo',
             stderr: 'Bar',
             stdoutEncoding: utf8,
-            stderrEncoding: utf8) // pub publish
+            stderrEncoding: utf8), // pub publish for plugin1
+        MockProcess(
+            stdout: 'Baz',
+            stdoutEncoding: utf8,
+            stderrEncoding: utf8), // pub publish for plugin1
       ];
 
-      final List<String> output = await runCapturingPrint(
-          commandRunner, <String>['publish-plugin', '--package', 'foo']);
+      final List<String> output = await runCapturingPrint(commandRunner,
+          <String>['publish-plugin', '--packages=plugin1,plugin2']);
 
       expect(
           output,
           containsAllInOrder(<Matcher>[
+            contains('Running `pub publish ` in /packages/plugin1...'),
             contains('Foo'),
             contains('Bar'),
+            contains('Package published!'),
+            contains('Running `pub publish ` in /packages/plugin2...'),
+            contains('Baz'),
+            contains('Package published!'),
           ]));
     });
 
@@ -172,7 +175,7 @@ void main() {
       mockStdin.mockUserInputs.add(utf8.encode('user input'));
 
       await runCapturingPrint(
-          commandRunner, <String>['publish-plugin', '--package', 'foo']);
+          commandRunner, <String>['publish-plugin', '--packages=foo']);
 
       expect(processRunner.mockPublishProcess.stdinMock.lines,
           contains('user input'));
@@ -184,17 +187,16 @@ void main() {
 
       await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
         '--pub-publish-flags',
-        '--dry-run,--server=foo'
+        '--dry-run,--server=bar'
       ]);
 
       expect(
           processRunner.recordedCalls,
           contains(ProcessCall(
               flutterCommand,
-              const <String>['pub', 'publish', '--dry-run', '--server=foo'],
+              const <String>['pub', 'publish', '--dry-run', '--server=bar'],
               pluginDir.path)));
     });
 
@@ -207,18 +209,17 @@ void main() {
 
       await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
         '--skip-confirmation',
         '--pub-publish-flags',
-        '--server=foo'
+        '--server=bar'
       ]);
 
       expect(
           processRunner.recordedCalls,
           contains(ProcessCall(
               flutterCommand,
-              const <String>['pub', 'publish', '--server=foo', '--force'],
+              const <String>['pub', 'publish', '--server=bar', '--force'],
               pluginDir.path)));
     });
 
@@ -233,8 +234,7 @@ void main() {
       final List<String> output =
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
       ], errorHandler: (Error e) {
         commandError = e;
       });
@@ -243,7 +243,7 @@ void main() {
       expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains('Publish foo failed.'),
+            contains('Publishing foo failed.'),
           ]));
     });
 
@@ -254,8 +254,7 @@ void main() {
       final List<String> output =
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
         '--dry-run',
       ]);
 
@@ -279,8 +278,7 @@ void main() {
       final List<String> output =
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        packageName,
+        '--packages=$packageName',
       ]);
 
       expect(
@@ -300,8 +298,7 @@ void main() {
       createFakePlugin('foo', packagesDir, examples: <String>[]);
       await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
       ]);
 
       expect(processRunner.recordedCalls,
@@ -319,8 +316,7 @@ void main() {
       final List<String> output =
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
       ], errorHandler: (Error e) {
         commandError = e;
       });
@@ -329,7 +325,7 @@ void main() {
       expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains('Publish foo failed.'),
+            contains('Publishing foo failed.'),
           ]));
       expect(
           processRunner.recordedCalls,
@@ -347,8 +343,7 @@ void main() {
       final List<String> output =
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
       ]);
 
       expect(
@@ -358,7 +353,7 @@ void main() {
       expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains('Released [foo] successfully.'),
+            contains('Published foo successfully.'),
           ]));
     });
 
@@ -371,8 +366,7 @@ void main() {
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
         '--skip-confirmation',
-        '--package',
-        'foo',
+        '--packages=foo',
       ]);
 
       expect(
@@ -382,7 +376,7 @@ void main() {
       expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains('Released [foo] successfully.'),
+            contains('Published foo successfully.'),
           ]));
     });
 
@@ -393,7 +387,7 @@ void main() {
       mockStdin.readLineOutput = 'y';
 
       final List<String> output = await runCapturingPrint(commandRunner,
-          <String>['publish-plugin', '--package', 'foo', '--dry-run']);
+          <String>['publish-plugin', '--packages=foo', '--dry-run']);
 
       expect(
           processRunner.recordedCalls
@@ -418,8 +412,7 @@ void main() {
       final List<String> output =
           await runCapturingPrint(commandRunner, <String>[
         'publish-plugin',
-        '--package',
-        'foo',
+        '--packages=foo',
         '--remote',
         'origin',
       ]);
@@ -431,42 +424,22 @@ void main() {
       expect(
           output,
           containsAllInOrder(<Matcher>[
-            contains('Released [foo] successfully.'),
+            contains('Published foo successfully.'),
           ]));
     });
   });
 
   group('Auto release (all-changed flag)', () {
     test('can release newly created plugins', () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>[],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>[],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       // Non-federated
       final Directory pluginDir1 = createFakePlugin('plugin1', packagesDir);
@@ -492,7 +465,7 @@ void main() {
             'Local repo is ready!',
             'Running `pub publish ` in ${pluginDir1.path}...\n',
             'Running `pub publish ` in ${pluginDir2.path}...\n',
-            'Packages released: plugin1, plugin2',
+            'Packages released: plugin1, plugin2/plugin2',
             'Done!'
           ]));
       expect(
@@ -507,42 +480,20 @@ void main() {
 
     test('can release newly created plugins, while there are existing plugins',
         () async {
-      const Map<String, dynamic> httpResponsePlugin0 = <String, dynamic>{
+      mockHttpResponses['plugin0'] = <String, dynamic>{
         'name': 'plugin0',
         'versions': <String>['0.0.1'],
       };
 
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>[],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>[],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin0.json') {
-          return http.Response(json.encode(httpResponsePlugin0), 200);
-        } else if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       // The existing plugin.
       createFakePlugin('plugin0', packagesDir);
@@ -575,7 +526,7 @@ void main() {
             'Local repo is ready!',
             'Running `pub publish ` in ${pluginDir1.path}...\n',
             'Running `pub publish ` in ${pluginDir2.path}...\n',
-            'Packages released: plugin1, plugin2',
+            'Packages released: plugin1, plugin2/plugin2',
             'Done!'
           ]));
       expect(
@@ -589,35 +540,16 @@ void main() {
     });
 
     test('can release newly created plugins, dry run', () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>[],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>[],
       };
 
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
       // Non-federated
       final Directory pluginDir1 = createFakePlugin('plugin1', packagesDir);
       // federated
@@ -651,7 +583,7 @@ void main() {
             'Running `pub publish ` in ${pluginDir2.path}...\n',
             'Tagging release plugin2-v0.0.1...',
             'Pushing tag to upstream...',
-            'Packages released: plugin1, plugin2',
+            'Packages released: plugin1, plugin2/plugin2',
             'Done!'
           ]));
       expect(
@@ -661,35 +593,15 @@ void main() {
     });
 
     test('version change triggers releases.', () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>['0.0.1'],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>['0.0.1'],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       // Non-federated
       final Directory pluginDir1 =
@@ -716,7 +628,7 @@ void main() {
             'Local repo is ready!',
             'Running `pub publish ` in ${pluginDir1.path}...\n',
             'Running `pub publish ` in ${pluginDir2.path}...\n',
-            'Packages released: plugin1, plugin2',
+            'Packages released: plugin1, plugin2/plugin2',
             'Done!'
           ]));
       expect(
@@ -732,35 +644,15 @@ void main() {
     test(
         'delete package will not trigger publish but exit the command successfully.',
         () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>['0.0.1'],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>['0.0.1'],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       // Non-federated
       final Directory pluginDir1 =
@@ -786,7 +678,7 @@ void main() {
             'Checking local repo...',
             'Local repo is ready!',
             'Running `pub publish ` in ${pluginDir1.path}...\n',
-            'The file at The pubspec file at ${pluginDir2.childFile('pubspec.yaml').path} does not exist. Publishing will not happen for plugin2.\nSafe to ignore if the package is deleted in this commit.\n',
+            'The pubspec file at ${pluginDir2.childFile('pubspec.yaml').path} does not exist. Publishing will not happen for plugin2.\nSafe to ignore if the package is deleted in this commit.\n',
             'Packages released: plugin1',
             'Done!'
           ]));
@@ -798,35 +690,15 @@ void main() {
 
     test('Existing versions do not trigger release, also prints out message.',
         () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>['0.0.2'],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>['0.0.2'],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       // Non-federated
       final Directory pluginDir1 =
@@ -871,35 +743,15 @@ void main() {
     test(
         'Existing versions do not trigger release, but fail if the tags do not exist.',
         () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'plugin1',
         'versions': <String>['0.0.2'],
       };
 
-      const Map<String, dynamic> httpResponsePlugin2 = <String, dynamic>{
+      mockHttpResponses['plugin2'] = <String, dynamic>{
         'name': 'plugin2',
         'versions': <String>['0.0.2'],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'plugin1.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        } else if (request.url.pathSegments.last == 'plugin2.json') {
-          return http.Response(json.encode(httpResponsePlugin2), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       // Non-federated
       final Directory pluginDir1 =
@@ -970,28 +822,10 @@ void main() {
     });
 
     test('Do not release flutter_plugin_tools', () async {
-      const Map<String, dynamic> httpResponsePlugin1 = <String, dynamic>{
+      mockHttpResponses['plugin1'] = <String, dynamic>{
         'name': 'flutter_plugin_tools',
         'versions': <String>[],
       };
-
-      final MockClient mockClient = MockClient((http.Request request) async {
-        if (request.url.pathSegments.last == 'flutter_plugin_tools.json') {
-          return http.Response(json.encode(httpResponsePlugin1), 200);
-        }
-        return http.Response('', 500);
-      });
-      final PublishPluginCommand command = PublishPluginCommand(packagesDir,
-          processRunner: processRunner,
-          stdinput: mockStdin,
-          httpClient: mockClient,
-          gitDir: gitDir);
-
-      commandRunner = CommandRunner<void>(
-        'publish_check_command',
-        'Test for publish-check command.',
-      );
-      commandRunner.addCommand(command);
 
       final Directory flutterPluginTools =
           createFakePlugin('flutter_plugin_tools', packagesDir);